Rename backend file_io examples and tidy fuzz/todo/kqueue
Rename per-backend file_io examples to uring_file_io, iocp_file_io, and kqueue_file_io for clarity. Minor fuzz/Cargo.toml, todo.md, and kqueue lib.rs cleanups.

+++ tpt-torus-backend-kqueue/src/lib.rs
@@ -670,7 +670,7 @@ impl Backend for KqueueBackend {
             }
         }
 
-        self.in_flight.fetch_add(submitted, Ordering::Relaxed);
+        self.in_flight.fetch_add(submitted as u32, Ordering::Relaxed);
         Ok(submitted as usize)
     }
